--- cse_hub/evaluation/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class EvaluationConfig(AppConfig):
-    name = 'evaluation'
